// src/models/dto/token-network-overview.dto.ts
export class TokenNetworkOverviewDto {
    month?: string;
    tokenNetwork?: string;
    channelOpened?: number;
    channelOpenedTot?: number;
    channelClosed?: number;
    channelClosedTot?: number;
    channelSettled?: number;
    channelSettledTot?: number;
    depositCount?: number
    depositCountTot?: number
    depositAmount?: number
    depositAmountTot?: number
    withdrawCount?: number
    withdrawCountTot?: number
    withdrawAmount?: number
    withdrawAmountTot?: number
}